feat(console): react to envfile changes (#6831)
Fixes #6349.
