+++ libs/wingsdk/package.json
@@ -104,6 +104,8 @@
     "constructs": "^10.3",
     "cron-parser": "^4.9.0",
     "cron-validator": "^1.3.1",
+    "dotenv": "^16.4.5",
+    "dotenv-expand": "^10.0.0",
     "express": "^4.19.2",
     "glob": "^8.1.0",
     "google-auth-library": "^8.9.0",
@@ -148,6 +150,8 @@
     "cdktf",
     "cron-parser",
     "cron-validator",
+    "dotenv",
+    "dotenv-expand",
     "express",
     "glob",
     "google-auth-library",
