+++ libs/wingsdk/src/helpers.ts
@@ -1,8 +1,11 @@
 // Code in this file will be automatically included in all inflight code bundles,
 // so avoid importing anything heavy here.
 import { notDeepStrictEqual } from "node:assert";
+import fs from "node:fs";
 import * as path from "node:path";
 import type { Construct } from "constructs";
+import { parse } from "dotenv";
+import { expand } from "dotenv-expand";
 import type { Resource } from "./std";
 import type { Node } from "./std/node";
 // since we moved from node:18 to node:20 the deepStrictEqual doesn't work as expected.
@@ -228,4 +231,48 @@ export function preflightClassSingleton(
     `${type.name}_singleton_${typeId}`
   );
   return root.resourceSingletons[type];
+}
+
+const DEFAULT_ENV_FILES = [`.env`, `.env.local`];
+
+/**
+ * Options for loading environment variables.
+ */
+export interface EnvLoadOptions {
+  /**
+   * The directory to load the environment variables from.
+   * @default process.cwd()
+   */
+  readonly cwd?: string;
+}
+
+/**
+ * Loads environment variables from `.env` and `.env.local` files.
+ */
+export function loadEnvVariables(
+  options?: EnvLoadOptions
+): Record<string, string> | undefined {
+  const envDir = options?.cwd ?? process.cwd();
+  const envFiles = DEFAULT_ENV_FILES.map((file) => path.join(envDir, file));
+
+  // Parse `envFiles` and combine their variables into a single object
+  const parsed = Object.fromEntries(
+    envFiles.flatMap((file) => {
+      try {
+        return Object.entries(parse(fs.readFileSync(file)));
+      } catch (_) {
+        return [];
+      }
+    })
+  );
+
+  // Expand and force load the environment variables
+  const expandedEnvVariables = expand({ parsed, ignoreProcessEnv: true });
+  if (expandedEnvVariables.parsed) {
+    for (const [key, value] of Object.entries(expandedEnvVariables.parsed)) {
+      process.env[key] = value;
+    }
+  }
+
+  return expandedEnvVariables.parsed;
 }
